+++ src/components/domain/AuthLoading.tsx
@@ -1,10 +1,11 @@
-import styled from "@emotion/styled";
 import { useEffect, useState } from "react";
+import styled from "@emotion/styled";
 import Lottie from "react-lottie";
+import { useAuthContext } from "@contexts/hooks";
 import * as animationData from "../../../public/assets/lottie/basketball.json";
 
-const width = 13;
-const fadeoutTimeMs = 400;
+const WIDTH = 13;
+const FADE_OUT_TIME_MS = 400;
 
 const defaultOptions = {
   loop: true,
@@ -15,30 +16,25 @@ const defaultOptions = {
   },
 };
 
-interface Props {
-  isShow: boolean;
-}
+const AuthLoading = () => {
+  const { authProps } = useAuthContext();
+  const { isLoading } = authProps;
 
-const AuthLoading = ({ isShow }: Props) => {
   const [isDisplay, setIsDisplay] = useState(true);
 
-  const getHeight = (width: number) => width * 1.32;
-
   useEffect(() => {
-    if (!isShow) {
-      setTimeout(() => setIsDisplay(false), fadeoutTimeMs);
+    if (isLoading) {
+      setIsDisplay(true);
+    } else {
+      setTimeout(() => setIsDisplay(false), FADE_OUT_TIME_MS);
     }
-  }, [isShow]);
+  }, [isLoading]);
 
   return isDisplay ? (
-    <LoadingContainer isShow={isShow}>
+    <LoadingContainer isShow={isLoading}>
       <LogoWrapper>
         <LogoText>Slam</LogoText>
-        <Lottie
-          options={defaultOptions}
-          height={getHeight(width)}
-          width={width}
-        />
+        <Lottie options={defaultOptions} width={WIDTH} />
       </LogoWrapper>
     </LoadingContainer>
   ) : (
@@ -59,7 +55,7 @@ const LoadingContainer = styled.div<{ isShow: boolean }>`
   backdrop-filter: blur(8px);
 
   // 애니메이션
-  transition: all ${fadeoutTimeMs}ms ease-out;
+  transition: all ${FADE_OUT_TIME_MS}ms ease-out;
   opacity: ${({ isShow }) => (isShow ? 1 : 0)};
 `;
 
@@ -73,7 +69,7 @@ const LogoText = styled.span`
   font-size: 30px;
   font-family: "Righteous", sans-serif;
   font-weight: 900;
-  margin-bottom: 6px;
+  margin-bottom: 13px;
   margin-right: 2px;
 `;
