+++ src/components/domain/index.ts
@@ -3,3 +3,4 @@ export { default as BottomNavigation } from "./BottomNavigtion";
 export { default as KakaoLogin } from "./KakaoLogin";
 export { default as DatePicker } from "./DatePicker";
 export { default as SlotPicker } from "./SlotPicker";
+export { default as AuthLoading } from "./AuthLoading";
