+++ src/components/layout/DefaultLayout/index.tsx
@@ -1,7 +1,7 @@
 import React, { ReactNode, useEffect, useRef, useState } from "react";
 import styled from "@emotion/styled";
-import { useNavigationContext } from "@contexts/NavigationProvider";
 import { BottomNavigation, TopNavigation } from "@components/domain";
+import { useNavigationContext } from "@contexts/hooks";
 import Container from "./Container";
 
 interface Props {
@@ -9,14 +9,13 @@ interface Props {
 }
 
 const DefaultLayout = ({ children }: Props) => {
+  const { navigationProps } = useNavigationContext();
+  const { isBottomNavigation, isTopNavigation } = navigationProps;
+
   const [isTransparent, setIsTransparent] = useState(true);
 
   const topNavigationRef = useRef<HTMLElement>(null);
   const containerRef = useRef<HTMLDivElement>(null);
-
-  const {
-    navigationProps: { isBottomNavigation, isTopNavigation },
-  } = useNavigationContext();
 
   const handleScroll = () =>
     topNavigationRef.current &&
