+++ src/contexts/hooks.ts
@@ -1,0 +1,2 @@
+export { default as useAuthContext } from "@contexts/AuthProvider/hook";
+export { default as useNavigationContext } from "@contexts/NavigationProvider/hook";
